+++ src/main/java/com/ahmetdayi/finalcase/entity/response/CreateClientResponse.java
@@ -6,7 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
+
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -30,7 +30,7 @@ public class CreateClientResponse {
 
     private LocalDate birthDay;
     @NotNull
-    private BigDecimal guarantee;
+    private double guarantee;
 
 
 }
